+++ src/config/entities/configuration.ts
@@ -157,6 +157,7 @@ export default () => ({
   },
   bridge: {
     baseUri: 'https://li.quest',
+    apiKey: process.env.BRIDGE_API_KEY,
   },
   contracts: {
     trustedForDelegateCall: {
