+++ src/domain/bridge/bridge.repository.interface.ts
@@ -1,9 +1,15 @@
 import type { BridgeName } from '@/domain/bridge/entities/bridge-name.entity';
-import type { RoutePreference } from '@/domain/bridge/entities/bridge-preference.entity';
 import type { BridgeQuote } from '@/domain/bridge/entities/bridge-quote.entity';
 import type { BridgeStatus } from '@/domain/bridge/entities/bridge-status.entity';
+import type {
+  AllowDenyPrefer,
+  RoutePreference,
+} from '@/domain/bridge/entities/bridge-preference.entity';
+import type {
+  BridgeRoute,
+  OrderType,
+} from '@/domain/bridge/entities/bridge-route.entity';
 import type { ExchangeName } from '@/domain/bridge/entities/exchange-name.entity';
-import type { OrderType } from '@/domain/bridge/entities/order-type.entity';
 import type { TimingStrategies } from '@/domain/bridge/entities/timing-strategies';
 
 export const IBridgeRepository = Symbol('IBridgeRepository');
@@ -44,4 +50,19 @@ export interface IBridgeRepository {
     routeTimingStrategies?: Array<TimingStrategies>;
     skipSimulation?: boolean;
   }): Promise<BridgeQuote>;
+
+  getRoutes(args: {
+    integrator?: string;
+    fee?: number;
+    maxPriceImpact?: number;
+    order?: OrderType;
+    slippage?: number;
+    referrer?: string;
+    allowSwitchChain?: boolean;
+    allowDestinationCall?: boolean;
+    bridges?: AllowDenyPrefer<BridgeName>;
+    exchanges?: AllowDenyPrefer<ExchangeName>;
+    swapStepTimingStrategies?: Array<TimingStrategies>;
+    routeTimingStrategies?: Array<TimingStrategies>;
+  }): Promise<Array<BridgeRoute>>;
 }
