--- src/domain/bridge/entities/order-type.entity.ts
@@ -1,5 +1,0 @@
-import { z } from 'zod';
-
-export const OrderTypeSchema = z.enum(['FASTEST', 'CHEAPEST']);
-
-export type OrderType = z.infer<typeof OrderTypeSchema>;
